+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `network_interceptor` to the `supertokens_config` in `init`.
   - This can be used to capture/modify all the HTTP requests sent to the core.
   - Solves the issue - https://github.com/supertokens/supertokens-core/issues/865
+- Functions `get_users_oldest_first`, `get_users_newest_first`, `get_user_count`, `delete_user`, `create_user_id_mapping`, `get_user_id_mapping`, `delete_user_id_mapping` and `update_or_delete_user_id_mapping_info` now accept `user_context` as an optional argument.
 - Fixed the dependencies in the example apps
   - Example apps will now fetch the latest version of the frameworks